LPD-26919 Rename package
