+++ modules/apps/commerce/commerce-service/src/main/java/com/liferay/commerce/internal/notification/term/provider/SalesAgentNotificationTermProvider.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.commerce.internal.notification.term.contributor;
+package com.liferay.commerce.internal.notification.term.provider;
 
 import com.liferay.notification.term.provider.NotificationTermProvider;
 import com.liferay.portal.kernel.util.HashMapBuilder;
