+++ modules/apps/notification/notification-service/src/main/java/com/liferay/notification/internal/term/provider/NotificationTermProviderRegistryImpl.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.notification.internal.term.contributor;
+package com.liferay.notification.internal.term.provider;
 
 import com.liferay.notification.term.provider.NotificationTermProvider;
 import com.liferay.notification.term.provider.NotificationTermProviderRegistry;
